Tests/LibWeb/Text/expected/css/PropertyOwningCSSStyleDeclaration-serialized-custom-properties.txt

test { --color: red; color: rgb(255, 0, 0); }
